requirements.txt
appdirs==1.4.3
asgiref==3.2.7
atomicwrites==1.4.0
attrs==19.3.0
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
distlib==0.3.0
Django==3.0.1
docutils==0.15.2
filelock==3.0.12
Flask==1.1.2
get==2019.4.13
gTTS==2.1.1
gTTS-token==1.1.3
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.6.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
Kivy==1.11.1
Kivy-Garden==0.1.4
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.3.3
more-itertools==8.3.0
mysql==0.0.2
mysql-connector==2.2.9
mysqlclient==1.4.6
numpy==1.19.4
packaging==20.4
Pillow==8.0.1
pluggy==0.13.1
post==2019.4.13
public==2019.4.13
py==1.8.1
Pygments==2.5.2
pyparsing==2.4.7
pytest==5.4.2
python-dateutil==2.8.1
pytz==2019.3
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
selenium==3.141.0
six==1.14.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.7
virtualenv==20.0.17
wcwidth==0.1.9
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
